+++ galsim/gsobject.py
@@ -1780,6 +1780,8 @@ class GSObject(object):
         #
 
         flux = self.flux
+        if flux == 0.0:
+            return 0, 1.0
         posflux = self.getPositiveFlux()
         negflux = self.getNegativeFlux()
         eta = negflux / (posflux + negflux)