+++ CHANGELOG.md
@@ -86,6 +86,9 @@ Deprecated Features
 - Deprecated ability to create multiple PhaseScreenPSFs with single call
   to makePSF, since it is now just as efficient to call makePSF multiple
   times. (#824)
+- Deprecated the use of `np.trapz` and `galsim.integ.mipdt` as valid
+  integration rules for use by `ImageIntegrator`s, replaced by
+  `galsim.integ.trapzRule` and `galsim.integ.midptRule`. (#887)
 - Changed the Angle.rad method to a property, so now you should write
   angle.rad rather than angle.rad() to get the value in radians.  Note: the
   return value is a subclass of float, so angle.rad() still works but gives a
@@ -228,6 +231,7 @@ New Features
 - Added `recenter` option to drawKImage to optionally not recenter the input
   image at (0,0).  The default `recenter=True` is consistent with how this
   function has worked in previous versions. (#873)
+- Drawing chromatic objects now uses significantly less RAM. (#887)
 
 
 New config features